move GrapesJsMailType from project-base to the framework package
