+++ packages/framework/src/Resources/views/Admin/Form/grapesJsMailType.html.twig
