+++ project-base/app/config/packages/twig.yaml
@@ -20,6 +20,8 @@ twig:
         - '@ShopsysFramework/Admin/Form/displayOnlyUrl.html.twig'
         - '@ShopsysFramework/Admin/Form/displayVariablesFormTheme.html.twig'
         - '@ShopsysFramework/Admin/Form/fileuploadFields.html.twig'
+        - '@ShopsysFramework/Admin/Form/grapesJsMailType.html.twig'
+        - '@ShopsysFramework/Admin/Form/grapesJsType.html.twig'
         - '@ShopsysFramework/Admin/Form/hreflangSettingFormTheme.html.twig'
         - '@ShopsysFramework/Admin/Form/imageuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/localizedFullWidth.html.twig'
@@ -39,8 +41,6 @@ twig:
         - '@ShopsysFramework/Admin/Form/rolesType.html.twig'
         - '@ShopsysFramework/Admin/Form/storeOpeningHoursFormTheme.html.twig'
         - '@ShopsysFramework/Admin/Form/transportPricesFormTheme.html.twig'
-        - 'Admin/Form/grapesJsMailType.html.twig'
-        - 'Admin/Form/grapesJsType.html.twig'
         - 'Admin/Form/languageConstantFormTheme.html.twig'
         - 'Admin/Form/productDetailFormTheme.html.twig'
         - 'Admin/Form/productVideoValue.html.twig'
